fix: implemented Mapper to map from custom type system to Terraform plugin protocol v5 block
Block definition now differentiate between attributes and nested blocks. Before that a nested block was also represented as attribute. Internally it is still an attribute but now we can implement two properties with same attribute name but one being an attribute and the other being a nested block.
Fixed a bug where the protocol version was not explicitly specified in TF_REATTACH_PROVIDERS environment variable.